Resolve naming errors and merge test scripts
Move upload log test from deleted test_server.rb to
test_remote_process.rb
